fw_unit: use the same type of 'request' parameter of ioctl(2)
The type of a second parameter of ioctl(2) is 'unsigned long', however
'hinawa_fw_unit_ioctl()' has 'int' type for this parameter.

This commit changes the type of argument.

+++ src/fw_unit.c
@@ -311,7 +311,8 @@ const guint8 *hinawa_fw_unit_get_config_rom(HinawaFwUnit *self, guint *length)
 }
 
 /* Internal use only. */
-void hinawa_fw_unit_ioctl(HinawaFwUnit *self, int req, void *args, int *err)
+void hinawa_fw_unit_ioctl(HinawaFwUnit *self, unsigned long req, void *args,
+			  int *err)
 {
 	HinawaFwUnitPrivate *priv;
